Verify the MD5 checksums of files following download. Upon failure, tell the user, delete the file and bail out so they can try again.
In addition, use the checksum to verify existing files in the download directory. When attempting to get a unique filename, if the name we generate exists, MD5sum it. If it matches, then just use that file. If it doesn't, move to the next filename and try again. If we get a name that hasn't been used before, then we obviously need to download.