Switch to using a joinable thread for downloads.
With newer versions of the macOS SDK, there's a race
condition in which the download thread classes destructor
is called before the status has been checked. This causes
a successful download to look like it failed.

This is a quick hack to fix the problem. We really need to
review the code properly and clean it up.